# src/Bokmal.Web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Bokmal.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5080/"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Bokmal.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/books/genres": {
      "get": {
        "tags": [
          "Books"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/books/top": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TopBookDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{slug}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/me": {
      "get": {
        "tags": [
          "Loans"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyLoansDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans": {
      "post": {
        "tags": [
          "Loans"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{loanId}/return": {
      "post": {
        "tags": [
          "Loans"
        ],
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/session": {
      "post": {
        "tags": [
          "Session"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BorrowerDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Session"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BorrowerDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/borrowers": {
      "get": {
        "tags": [
          "Session"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BorrowerDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AvailabilityDto": {
        "required": [
          "totalCopies",
          "availableCopies",
          "onLoanCopies",
          "isAvailable"
        ],
        "type": "object",
        "properties": {
          "totalCopies": {
            "type": "integer",
            "format": "int32"
          },
          "availableCopies": {
            "type": "integer",
            "format": "int32"
          },
          "onLoanCopies": {
            "type": "integer",
            "format": "int32"
          },
          "isAvailable": {
            "type": "boolean"
          }
        }
      },
      "BookDetailDto": {
        "required": [
          "id",
          "slug",
          "title",
          "author",
          "genre",
          "publishedYear",
          "pageCount",
          "description",
          "availability",
          "readingTime",
          "alsoBorrowed"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "publishedYear": {
            "type": "integer",
            "format": "int32"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "availability": {
            "$ref": "#/components/schemas/AvailabilityDto"
          },
          "readingTime": {
            "$ref": "#/components/schemas/ReadingTimeDto"
          },
          "alsoBorrowed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationDto"
            }
          }
        }
      },
      "BookSummaryDto": {
        "required": [
          "id",
          "slug",
          "title",
          "author",
          "genre",
          "publishedYear",
          "pageCount",
          "availability",
          "readingTime"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "genre": {
            "type": "string"
          },
          "publishedYear": {
            "type": "integer",
            "format": "int32"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32"
          },
          "availability": {
            "$ref": "#/components/schemas/AvailabilityDto"
          },
          "readingTime": {
            "$ref": "#/components/schemas/ReadingTimeDto"
          }
        }
      },
      "BorrowerDto": {
        "required": [
          "id",
          "email",
          "displayName",
          "joinedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BorrowRequest": {
        "required": [
          "bookSlug"
        ],
        "type": "object",
        "properties": {
          "bookSlug": {
            "type": "string"
          }
        }
      },
      "LoanDto": {
        "required": [
          "id",
          "bookId",
          "bookSlug",
          "bookTitle",
          "author",
          "copyNumber",
          "borrowedAt",
          "dueAt",
          "returnedAt",
          "isOverdue"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookId": {
            "type": "string",
            "format": "uuid"
          },
          "bookSlug": {
            "type": "string"
          },
          "bookTitle": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "copyNumber": {
            "type": "integer",
            "format": "int32"
          },
          "borrowedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          },
          "returnedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isOverdue": {
            "type": "boolean"
          }
        }
      },
      "MyLoansDto": {
        "required": [
          "current",
          "past"
        ],
        "type": "object",
        "properties": {
          "current": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanDto"
            }
          },
          "past": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanDto"
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ReadingTimeDto": {
        "required": [
          "typicalDays",
          "basedOnLoans",
          "fromHistory"
        ],
        "type": "object",
        "properties": {
          "typicalDays": {
            "type": "integer",
            "format": "int32"
          },
          "basedOnLoans": {
            "type": "integer",
            "format": "int32"
          },
          "fromHistory": {
            "type": "boolean"
          }
        }
      },
      "RecommendationDto": {
        "required": [
          "book",
          "sharedBorrowers"
        ],
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/BookSummaryDto"
          },
          "sharedBorrowers": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SignInRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "TopBookDto": {
        "required": [
          "book",
          "borrowCount"
        ],
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/BookSummaryDto"
          },
          "borrowCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Bokmal.Api"
    },
    {
      "name": "Books"
    },
    {
      "name": "Loans"
    },
    {
      "name": "Session"
    }
  ]
}